+++ InkbunnyLib/LoginResponse.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace InkbunnyLib {
 	public class LoginResponse : InkbunnyResponse {
 		public string sid { get; set; }
-		public long user_id { get; set; }
+		public int user_id { get; set; }
 		public string ratingsmask { get; set; }
 	}
 }
